requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
atk-1.0=2.36.0=h3371d22_4
blas=1.0=mkl
bottleneck=1.3.2=py38heb32a55_1
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
cairo=1.16.0=ha12eb4b_1010
certifi=2021.10.8=py38h578d9bd_1
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=36.0.0=py38h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
dowhy=0.7=pypi_0
expat=2.4.7=h27087fc_0
ffmpeg=4.3=hf484d3e_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.13.96=h8e229c2_2
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.30.0=pypi_0
freetype=2.11.0=h70c0345_0
fribidi=1.0.10=h36c2ea0_0
gdk-pixbuf=2.42.6=h04a7f16_0
gettext=0.19.8.1=h73d1719_1008
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
graphite2=1.3.13=h58526e2_1001
graphviz=3.0.0=h5abf519_1
gtk2=2.24.33=h90689f9_2
gts=0.7.6=h64030ff_2
h5py=3.6.0=pypi_0
harfbuzz=3.4.0=hb4a5f5f_0
icu=69.1=h9c3ff4c_0
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
kiwisolver=1.4.0=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_14
libgd=2.3.3=h695aa2c_0
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglib=2.70.2=h174f98d_4
libiconv=1.16=h516909a_0
libidn2=2.3.2=h7f8727e_0
libnsl=2.0.0=h7f98852_0
libpng=1.6.37=hbc83047_0
librsvg=2.52.5=h0a9e6e8_2
libstdcxx-ng=11.2.0=he4da1e4_14
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libtool=2.4.6=h9c3ff4c_1008
libunistring=0.9.10=h27cfd23_0
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.13=h7f98852_1004
libxml2=2.9.12=h885dcf4_1
libzlib=1.2.11=h36c2ea0_1013
llvm-openmp=13.0.1=he0ac6c6_1
lz4-c=1.9.3=h295c915_1
matplotlib=3.5.1=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mpmath=1.2.1=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.7.1=pypi_0
numexpr=2.8.1=py38h6abb31d_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.0=h7f98852_2
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py38h295c915_0
pango=1.50.5=h4dcc4a0_1
patsy=0.5.2=pypi_0
pcre=8.45=h9c3ff4c_0
pillow=9.0.1=py38h22f2fdc_0
pip=21.2.4=py38h06a4308_0
pixman=0.40.0=h36c2ea0_0
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.21=pyhd3eb1b0_0
pydot=1.4.2=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h0744224_3_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-graphviz=0.19.1=pyhaef67bd_0
python_abi=3.8=2_cp38
pytorch=1.11.0=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=py38hc147768_0
seaborn=0.11.2=pypi_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.0=hc218d9a_0
statsmodels=0.13.2=pypi_0
sympy=1.10=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.11.0=py38_cu113
torchvision=0.12.0=py38_cu113
typing_extensions=3.10.0.2=pyh06a4308_0
umnn=1.68=pypi_0
urllib3=1.26.8=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.4.9=haebb681_0
